MCP tools refactor
