+++ mcp_tools/binary_analysis/binary_analysis_and_reverse_engineering.py
