+++ mcp_tools/binary_analysis/enhanced_binary_analysis_and_exploitation.py
