+++ mcp_tools/ctf_forensics/advanced_ctf_tools.py
