+++ mcp_tools/ops/system_monitoring.py
