+++ mcp_tools/ops/visual_output_tools.py
